# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/posts/feed": {
      "get": {
        "operationId": "PostController_feed",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedPostsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "posts"
        ]
      }
    },
    "/posts/{id}": {
      "get": {
        "operationId": "PostController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "posts"
        ]
      },
      "delete": {
        "operationId": "PostController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "posts"
        ]
      }
    },
    "/posts": {
      "post": {
        "operationId": "PostController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "posts"
        ]
      },
      "get": {
        "operationId": "PostController_list",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "groupId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "visibility",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedPostsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "posts"
        ]
      }
    },
    "/posts/{id}/like": {
      "post": {
        "operationId": "PostController_like",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "posts"
        ]
      },
      "delete": {
        "operationId": "PostController_unlike",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "posts"
        ]
      }
    },
    "/posts/{id}/comments": {
      "post": {
        "operationId": "PostController_addComment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "posts"
        ]
      },
      "get": {
        "operationId": "PostController_listComments",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedCommentsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "posts"
        ]
      }
    },
...
    "/notifications": {
      "get": {
        "operationId": "NotificationController_list",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "unreadOnly",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedNotificationsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "notifications"
        ]
      }
    },
    "/notifications/{id}/read": {
      "patch": {
        "operationId": "NotificationController_markRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "notifications"
        ]
      }
    },
...
    "/gamification/rewards": {
      "get": {
        "operationId": "GamificationController_listMyRewardEvents",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedRewardEventsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "gamification"
        ]
      }
    },
...
  },
  "info": {
    "title": "Habit Tracker API",
    "description": "API for habit tracking, community, challenges and gamification",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PostResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "groupId": {
            "type": "object"
          },
          "challengeId": {
            "type": "object"
          },
          "content": {
            "type": "object"
          },
          "type": {
            "type": "string",
            "enum": [
              "TEXT",
              "ACHIEVEMENT",
              "CHALLENGE_CREATED",
              "CHALLENGE_ON_TRACK",
              "CHALLENGE_COMPLETED",
              "IMAGE"
            ]
          },
          "mediaUrl": {
            "type": "object"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "PRIVATE"
            ]
          },
          "isAutoGenerated": {
            "type": "boolean"
          },
          "challengeTitle": {
            "type": "object"
          },
          "taggedUserIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "hashtags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "commentCount": {
            "type": "number"
          },
          "likeCount": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "userId",
          "type",
          "visibility",
          "isAutoGenerated",
          "taggedUserIds",
          "hashtags",
          "commentCount",
          "createdAt",
          "updatedAt"
        ]
      },
      "PaginatedPostsResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PostResponseDto"
            }
          },
          "nextCursor": {
            "type": "string"
          }
        },
        "required": [
          "items"
        ]
      },
      "AddCommentDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "maxLength": 5000,
            "example": "Great progress!"
          }
        },
        "required": [
          "content"
        ]
      },
      "CommentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "postId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "postId",
          "userId",
          "content",
          "createdAt"
        ]
      },
      "PaginatedCommentsResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommentResponseDto"
            }
          },
          "nextCursor": {
            "type": "string"
          }
        },
        "required": [
          "items"
        ]
      },
...
      "NotificationResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "receiverId": {
            "type": "string"
          },
          "senderId": {
            "type": "string"
          },
          "notificationTypeId": {
            "type": "string"
          },
          "senderName": {
            "type": "string"
          },
          "senderAvatar": {
            "type": "object"
          },
          "title": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "data": {
            "type": "object"
          },
          "isRead": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "receiverId",
          "senderId",
          "notificationTypeId",
          "senderName",
          "title",
          "body",
          "data",
          "isRead",
          "createdAt"
        ]
      },
      "PaginatedNotificationsResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NotificationResponseDto"
            }
          },
          "nextCursor": {
            "type": "string"
          }
        },
        "required": [
          "items"
        ]
      },
      "RewardEventResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "achievementDefinitionId": {
            "type": "string"
          },
          "pointsAwarded": {
            "type": "number"
          },
          "sourceType": {
            "type": "object"
          },
          "sourceId": {
            "type": "object"
          },
          "title": {
            "type": "object"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "userId",
          "achievementDefinitionId",
          "pointsAwarded",
          "createdAt"
        ]
      },
      "PaginatedRewardEventsResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RewardEventResponseDto"
            }
          },
          "nextCursor": {
            "type": "string"
          }
        },
        "required": [
          "items"
        ]
      },
...
    }
  }
}